requirements.txt
selenium
crewai
crewai-tools
sentence-transformers
chromadb
google-generativeai